+++ src/yafowil/tests/__init__.py
@@ -60,6 +60,12 @@ def test_suite():
     return suite
 
 
+def run_tests():
+    from zope.testrunner.runner import Runner
+
+    runner = Runner(found_suites=[test_suite()])
+    runner.run()
+
+
 if __name__ == '__main__':
-    runner = unittest.TextTestRunner(failfast=True)
-    runner.run(test_suite())
+    run_tests()